+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/mining/keego/Keego.java
@@ -50,8 +50,8 @@ public class Keego extends Feature {
 	public static final TagKey<Item> KEEGO_HAND_EQUIPMENT = TagKey.create(Registries.ITEM, new ResourceLocation(IguanaTweaksExpanded.MOD_ID, "equipment/hand/keego"));
 	public static final TagKey<Item> KEEGO_ARMOR_EQUIPMENT = TagKey.create(Registries.ITEM, new ResourceLocation(IguanaTweaksExpanded.MOD_ID, "equipment/armor/keego"));
 
-	public static final RegistryObject<MobEffect> MOVEMENT_MOMENTUM = ITERegistries.MOB_EFFECTS.register("movement_momentum", () -> new ILMobEffect(MobEffectCategory.BENEFICIAL, 0xFCD373, false).addAttributeModifier(Attributes.MOVEMENT_SPEED, "544cf3ee-676f-4685-aec7-a6b3d64875b0", 0.05d, AttributeModifier.Operation.MULTIPLY_BASE));
-	public static final RegistryObject<MobEffect> ATTACK_MOMENTUM = ITERegistries.MOB_EFFECTS.register("attack_momentum", () -> new ILMobEffect(MobEffectCategory.BENEFICIAL, 0xFCD373, false).addAttributeModifier(Attributes.ATTACK_SPEED, "f6fe8408-b88c-4e51-8892-8b20574cfc49", 0.05d, AttributeModifier.Operation.ADDITION));
+	public static final RegistryObject<MobEffect> MOVEMENT_MOMENTUM = ITERegistries.MOB_EFFECTS.register("movement_momentum", () -> new ILMobEffect(MobEffectCategory.BENEFICIAL, 0xFCD373, false).addAttributeModifier(Attributes.MOVEMENT_SPEED, "544cf3ee-676f-4685-aec7-a6b3d64875b0", 0.06d, AttributeModifier.Operation.MULTIPLY_BASE));
+	public static final RegistryObject<MobEffect> ATTACK_MOMENTUM = ITERegistries.MOB_EFFECTS.register("attack_momentum", () -> new ILMobEffect(MobEffectCategory.BENEFICIAL, 0xFCD373, false).addAttributeModifier(Attributes.ATTACK_SPEED, "f6fe8408-b88c-4e51-8892-8b20574cfc49", 0.06d, AttributeModifier.Operation.ADDITION));
 	public static final RegistryObject<MobEffect> MINING_MOMENTUM = ITERegistries.MOB_EFFECTS.register("mining_momentum", () -> new ILMobEffect(MobEffectCategory.BENEFICIAL, 0xFCD373, false));
 
 	public static final SimpleBlockWithItem ORE = SimpleBlockWithItem.register("keego_ore", () -> new KeegoOreBlock(BlockBehaviour.Properties.copy(Blocks.BEDROCK).strength(-1f, 10f), UniformInt.of(10, 15)));
@@ -96,7 +96,7 @@ public class Keego extends Feature {
 
 		//noinspection DataFlowIssue
 		int lvl = event.getEntity().getEffect(MINING_MOMENTUM.get()).getAmplifier() + 1;
-		event.setNewSpeed(event.getNewSpeed() * (1 + lvl * 0.05f));
+		event.setNewSpeed(event.getNewSpeed() * (1 + lvl * 0.10f));
 	}
 
 	@SubscribeEvent
@@ -110,7 +110,7 @@ public class Keego extends Feature {
 			//noinspection DataFlowIssue
 			amplifier = event.getPlayer().getEffect(MINING_MOMENTUM.get()).getAmplifier() + 1;
 
-		int duration = (int) (1f / event.getState().getDestroyProgress(event.getPlayer(), event.getLevel(), event.getPos()) + 5) * 3;
+		int duration = (int) (1f / event.getState().getDestroyProgress(event.getPlayer(), event.getLevel(), event.getPos()) + 5) * 3 + 1;
 		event.getPlayer().addEffect(new MobEffectInstance(MINING_MOMENTUM.get(), Math.max(duration, 15), Math.min(amplifier, 31), false, false, true));
 	}
 
@@ -122,20 +122,20 @@ public class Keego extends Feature {
 				|| event.phase == TickEvent.Phase.END)
 			return;
 
-		AtomicInteger maxAmplifier = new AtomicInteger(0);
+		AtomicInteger pieces = new AtomicInteger(0);
 		event.player.getInventory().armor.forEach(stack -> {
 			if (stack.is(KEEGO_ARMOR_EQUIPMENT))
-				maxAmplifier.addAndGet(2);
+				pieces.addAndGet(1);
 		});
-		if (maxAmplifier.get() == 0)
+		if (pieces.get() == 0)
 			return;
-		if (event.player.walkDist % 8 < event.player.walkDistO % 8) {
+		if (event.player.walkDist % (10 - pieces.get() * 2) < event.player.walkDistO % (10 - pieces.get() * 2)) {
 			int amplifier = 0;
 			if (event.player.hasEffect(MOVEMENT_MOMENTUM.get()))
 				//noinspection DataFlowIssue
 				amplifier = event.player.getEffect(MOVEMENT_MOMENTUM.get()).getAmplifier() + 1;
 
-			event.player.addEffect(new MobEffectInstance(MOVEMENT_MOMENTUM.get(), 100, Math.min(amplifier, maxAmplifier.get() - 1), false, false, true));
+			event.player.addEffect(new MobEffectInstance(MOVEMENT_MOMENTUM.get(), 100, Math.min(amplifier, 7), false, false, true));
 		}
 
 	}